+++ src/containers/InboxPage/InboxPage.js
@@ -14,6 +14,7 @@ import {
   LINE_ITEM_NIGHT,
   LINE_ITEM_HOUR,
   LISTING_UNIT_TYPES,
+  STOCK_MULTIPLE_ITEMS,
 } from '../../util/types';
 import { subtractTime } from '../../util/dates';
 import {
@@ -114,7 +115,14 @@ BookingTimeInfoMaybe.propTypes = {
 };
 
 export const InboxItem = props => {
-  const { transactionRole, tx, intl, stateData, isBooking, stockType = 'multipleItems' } = props;
+  const {
+    transactionRole,
+    tx,
+    intl,
+    stateData,
+    isBooking,
+    stockType = STOCK_MULTIPLE_ITEMS,
+  } = props;
   const { customer, provider, listing } = tx;
   const { processName, processState, actionNeeded, isSaleNotification, isFinal } = stateData;
   const isCustomer = transactionRole === TX_TRANSITION_ACTOR_CUSTOMER;
@@ -123,7 +131,7 @@ export const InboxItem = props => {
   const hasPricingData = lineItems.length > 0;
   const unitLineItem = getUnitLineItem(lineItems);
   const quantity = hasPricingData && !isBooking ? unitLineItem.quantity.toString() : null;
-  const showStock = stockType === 'multipleItems' || (quantity && unitLineItem.quantity > 1);
+  const showStock = stockType === STOCK_MULTIPLE_ITEMS || (quantity && unitLineItem.quantity > 1);
 
   const otherUser = isCustomer ? provider : customer;
   const otherUserDisplayName = <UserDisplayName user={otherUser} intl={intl} />;